modify types

+++ test/test_http/http_server.cpp
@@ -7,12 +7,12 @@ void on_new_request(http::connection_wptr &wconn, http::request_sptr &&req) {
     res.set_status_code(200);
     res.set_http_version(http::VERSION_11);
     res.get_header()->set("Content-Type", "text/html; charset=utf-8");
-    res.get_header()->set("Content-Length", data.size());
+    res.get_header()->set("Content-Length", (int32_t)data.size());
     // res.get_header()->set("Transfer-Encoding", "chunked");
     // conn->send(&res);
 
     http::content_sptr content(new http::content);
-    content->append(data.c_str(), data.size());
+    content->append(data.c_str(), (int32_t)data.size());
     res.set_content(content);
 
     auto conn = wconn.lock();
